requirements.txt
PyJWT[crypto]
requests
python-dotenv